--- benchmarks/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "adv-zlib-benchmarks",
-  "version": "1.0.0",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "start": "node inde.js"
-  },
-  "dependencies": {
-    "archiver": "^6.0.1"
-  }
-}
