+++ src/index.ts
@@ -1,3 +1,14 @@
 import { AdvZlib } from './adv-zlib.js';
-export type { ZipEntry } from './entry.js';
+
+export type { 
+  ZipEntry, 
+  Logger, 
+  AdvZlibOptions,
+  CentralDir,
+  CentralDirFileHeader,
+  LocalFileHeader,
+  EndOfCentralDirRecord,
+  Zip64EndOfCentralDirRecord
+} from './adv-zlib.js';
+
 export default AdvZlib;
