--- src/types.ts
@@ -1,11 +1,0 @@
-export interface Logger {
-  debug: (message: string, ...args: any[]) => void;
-  info: (message: string, ...args: any[]) => void;
-  warn: (message: string, ...args: any[]) => void;
-  error: (message: string, ...args: any[]) => void;
-}
-
-export interface MemoryUsage {
-  heapUsed: number;
-  external: number;
-}
