+++ tests/legacy-snapshot.test.ts
